move util, create stream manager folder
