+++ packages/core/src/lib/base_protocol.ts
@@ -14,7 +14,7 @@ import {
 } from "@waku/utils/libp2p";
 
 import { filterPeersByDiscovery } from "./filterPeers.js";
-import { StreamManager } from "./stream_manager.js";
+import { StreamManager } from "./stream_manager/index.js";
 
 /**
  * A class with predefined helpers, to be used as a base to implement Waku
