+++ packages/core/src/lib/stream_manager/index.ts
@@ -1,0 +1,1 @@
+export { StreamManager } from "./stream_manager.js";
